create contantMainTableWrap
