+++ src/index.js
@@ -1,7 +1,7 @@
 // Get components
 import headComponent from "./components/headComponent.js";
 import footer from "./components/get-footer.js";
-import ViewInBrowsLink from "./miscellaneous/view-in-browser-link.js";
+// import ViewInBrowsLink from "./miscellaneous/view-in-browser-link.js";
 import UnsubLink from "./miscellaneous/get-unsub-link.js";
 import recipe from "./components/get-recipe.js";
 import button from "./components/get-button.js";
@@ -13,8 +13,9 @@ import content from "./components/get-content.js";
 import tableForNotes from "./components/tableForNotes.js";
 import tableWithTwoRecipes from "./components/tableWithTwoRecipes.js";
 
-import topPanelViewInBrows from "./outerTemplate/topPanelViewInBrows.js";
-import tableHeaderBlockLogo from "./outerTemplate/tableHeaderBlockLogo.js";
+// import topPanelViewInBrows from "./outerTemplate/topPanelViewInBrows.js";
+// import tableHeaderBlockLogo from "./outerTemplate/tableHeaderBlockLogo.js";
+import contentMainTableWrap from "./outerTemplate/contentMainTableWrap.js";
 
 // typography start
 import headings from "../components/typography/src/components/get-headings.js";
@@ -25,7 +26,7 @@ import strongText from "../components/typography/src/components/strong.js";
 export {
   headComponent,
   footer,
-  ViewInBrowsLink,
+  // ViewInBrowsLink,
   UnsubLink,
   recipe,
   button,
@@ -36,8 +37,9 @@ export {
   content,
   tableForNotes,
   tableWithTwoRecipes,
-  topPanelViewInBrows,
-  tableHeaderBlockLogo,
+  // topPanelViewInBrows,
+  // tableHeaderBlockLogo,
+  contentMainTableWrap,
   headings,
   paragraph,
   strongText,
